Integrate cybersecurity Sigma rule generation environment
